Backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dotenv==1.1.0
dj-database-url==3.1.2
whitenoise==6.11.0
psycopg[binary]==3.3.3
